Removing the function caclulate_scanner_options.
This was a little bit confusing, as to waht is really caclulated. All
options (except n_pixels in some cases) hvae to be set explicitely.

+++ src/2d/barrel/lm_reconstruction_cmd.cpp
@@ -18,7 +18,7 @@
 #endif
 
 using F = float;
-using S = int;
+using S = short;
 using RNG = std::mt19937;
 using Point = PET2D::Point<F>;
